feat: 添加修改服务器信息

+++ src/components/server/Servers.vue
@@ -2,13 +2,13 @@
   <div>
     <el-breadcrumb separator="/">
       <el-breadcrumb-item :to="{ path: '/home' }">首页</el-breadcrumb-item>
-      <el-breadcrumb-item>服务器管理</el-breadcrumb-item>
-      <el-breadcrumb-item>服务器列表</el-breadcrumb-item>
+      <el-breadcrumb-item>节点管理</el-breadcrumb-item>
+      <el-breadcrumb-item>节点列表</el-breadcrumb-item>
     </el-breadcrumb>
 
     <el-card class="box-card">
       <div slot="header" class="clearfix">
-        <span>服务器列表</span>
+        <span>节点列表</span>
       </div>
       <el-row :gutter="20">
         <el-col :span="8">
@@ -18,8 +18,9 @@
         </el-col>
         <el-col :span="4">
           <router-link to="/servers/create">
-            <el-button type="primary" @click="createServer()">添加服务器</el-button>
+            <el-button type="primary" @click="createServer()">添加节点</el-button>
           </router-link>
+          <el-button class="refresh" type="primary" @click="refresh()">刷新节点</el-button>
         </el-col>
       </el-row>
       <el-table
@@ -61,7 +62,7 @@
         <el-table-column
             label="操作">
           <template slot-scope="scope">
-            <!--            <el-button size="mini" type="primary" icon="el-icon-edit" @click="editUser(scope.row._id)"></el-button>-->
+            <el-button size="mini" type="primary" icon="el-icon-edit" @click="editServer(scope.row._id)"></el-button>
             <el-button size="mini" type="danger" icon="el-icon-delete" @click="deleteServer(scope.row)"></el-button>
           </template>
         </el-table-column>
@@ -125,16 +126,50 @@ export default {
       window.sessionStorage.setItem("activePath", '/servers/create');
     },
     async deleteServer(serverInfo) {
-      await this.$http.delete(`/api/v2/servers/${serverInfo._id}`)
-          .then(response => {
-            const res = response.data;
-            if (res.meta.status !== 200) {
-              return this.$message.error("删除失败!");
-            }
-            this.$message.success("删除成功!");
-          })
-      await this.getServerList();
+      this.$confirm(`此操作将永久删除该节点 ${serverInfo.appName}, 是否继续？`, '提示', {
+        type: 'warning',
+        distinguishCancelAndClose: true,
+        confirmButtonText: '删除',
+        cancelButtonText: '取消'
+      }).then(async () => {
+        await this.$http.delete(`/api/v2/servers/${serverInfo._id}`)
+            .then(response => {
+              const res = response.data;
+              if (res.meta.status !== 200) {
+                return this.$message.error("删除失败!");
+              }
+              this.$message.success("删除成功!");
+            })
+        await this.getServerList();
+      }).catch(action => {
+        this.$message({
+          type: 'info',
+          message: action === 'cancel'
+              ? '放弃删除'
+              : '停留在当前页面'
+        })
+      });
+
     },
+    async editServer(nodeId) {
+      await this.$router.push(`/servers/${nodeId}/edit`);
+      window.sessionStorage.removeItem("activePath")
+    },
+    async refresh() {
+      await this.$http.get(`/api/v2/servers/refresh`)
+          .then(response => {
+            if (response) {
+              const res = response.data
+              if (res.meta.status === 200) {
+                this.$message.success("刷新节点信息成功!");
+              }
+            } else {
+              return this.$message.error("刷新节点信息失败!");
+            }
+          }).catch(e => {
+            this.$message.error(e.toString());
+          });
+    }
   }
 }
 </script>
@@ -146,5 +181,9 @@ export default {
 
 .el-pagination {
   margin-top: 15px;
+}
+
+.refresh {
+  margin-left: 15px;
 }
 </style>
